requirements.txt
uvicorn==0.34.0
fastapi==0.115.6
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1